# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "API 4",
        "title": "Mikroserwis zarządzania biletami i wydarzeniami - REST API",
        "contact": {
            "name": "Asia Kolis"
        },
        "license": {
            "name": "API License URL"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
    "tags": [
        {
            "name": "event-endpoint",
            "description": "Event Endpoint"
        },
        {
            "name": "ticket-endpoint",
            "description": "Ticket Endpoint"
        }
    ],
    "paths": {
        "/events/": {
            "post": {
                "tags": [
                    "event-endpoint"
                ],
                "summary": "createEventAndTickets",
                "operationId": "createEventAndTicketsUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "event",
                        "description": "event",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/events/resigantion/{eventid}": {
            "get": {
                "tags": [
                    "event-endpoint"
                ],
                "summary": "checkResignationTime",
                "operationId": "checkResignationTimeUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "eventid",
                        "in": "path",
                        "description": "eventid",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/events/seats/{eventid}": {
            "get": {
                "tags": [
                    "event-endpoint"
                ],
                "summary": "showAvailableSeatsNumber",
                "operationId": "showAvailableSeatsNumberUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventid",
                        "in": "path",
                        "description": "eventid",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/events/show": {
            "get": {
                "tags": [
                    "event-endpoint"
                ],
                "summary": "showEvents",
                "operationId": "showEventsUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "event",
                        "description": "event",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/events/user/{userid}": {
            "get": {
                "tags": [
                    "event-endpoint"
                ],
                "summary": "showEventsCreatedByUser",
                "operationId": "showEventsCreatedByUserUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "userid",
                        "in": "path",
                        "description": "userid",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/events/{eventid}": {
            "get": {
                "tags": [
                    "event-endpoint"
                ],
                "summary": "showEventInfo",
                "operationId": "showEventInfoUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventid",
                        "in": "path",
                        "description": "eventid",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "event-endpoint"
                ],
                "summary": "cancelEvent",
                "operationId": "cancelEventUsingDELETE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventid",
                        "in": "path",
                        "description": "eventid",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/events/{eventid}/status/{status}": {
            "put": {
                "tags": [
                    "event-endpoint"
                ],
                "summary": "updateEventStatus",
                "operationId": "updateEventStatusUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventid",
                        "in": "path",
                        "description": "eventid",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "description": "status",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/tickets/": {
            "post": {
                "tags": [
                    "ticket-endpoint"
                ],
                "summary": "addTicket",
                "operationId": "addTicketUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "ticket",
                        "description": "ticket",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Ticket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/tickets/event/{eventid}": {
            "delete": {
                "tags": [
                    "ticket-endpoint"
                ],
                "summary": "cancelTicketForEvent",
                "operationId": "cancelTicketForEventUsingDELETE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventid",
                        "in": "path",
                        "description": "eventid",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/tickets/event/{eventid}/available": {
            "get": {
                "tags": [
                    "ticket-endpoint"
                ],
                "summary": "showAvailableTickets",
                "operationId": "showAvailableTicketsUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventid",
                        "in": "path",
                        "description": "eventid",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Ticket"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/tickets/event/{eventid}/{regular},{premium}": {
            "post": {
                "tags": [
                    "ticket-endpoint"
                ],
                "summary": "addMoreTickets",
                "operationId": "addMoreTicketsUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventid",
                        "in": "path",
                        "description": "eventid",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "regular",
                        "in": "path",
                        "description": "regular",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "premium",
                        "in": "path",
                        "description": "premium",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/tickets/{ticketid}": {
            "get": {
                "tags": [
                    "ticket-endpoint"
                ],
                "summary": "showTicket",
                "operationId": "showTicketUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "ticketid",
                        "in": "path",
                        "description": "ticketid",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Ticket"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/tickets/{ticketid}/status/{status}": {
            "put": {
                "tags": [
                    "ticket-endpoint"
                ],
                "summary": "updateTicketStatus",
                "operationId": "updateTicketStatusUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "ticketid",
                        "in": "path",
                        "description": "ticketid",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "description": "status",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "Event": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "date": {
                    "type": "string",
                    "format": "date-time"
                },
                "eventAddress": {
                    "type": "string"
                },
                "eventType": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "organizer": {
                    "type": "string"
                },
                "premiumTicketsNumber": {
                    "type": "integer",
                    "format": "int32"
                },
                "regularTicketsNumber": {
                    "type": "integer",
                    "format": "int32"
                },
                "resignationPeriod": {
                    "type": "integer",
                    "format": "int32"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "Response": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        },
        "Ticket": {
            "type": "object",
            "properties": {
                "eventID": {
                    "type": "integer",
                    "format": "int64"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "place": {
                    "type": "string"
                },
                "price": {
                    "type": "integer",
                    "format": "int32"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}